refactor: consolidate in-memory loader impls
