+++ playground/wasm/Cargo.toml
@@ -14,7 +14,6 @@ wasm-bindgen = "0.2"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 console_error_panic_hook = "0.1"
-rustc-hash = "2"
 ecow = "0.2"
 
 lisette-format      = { path = "../../crates/format" }
